Upload TUF metadata in reverse read order to avoid errors
If a timestamp is downloaded before a snapshot is uploaded
for example, then there will be a msimatch in metadata. Once
consistent snapshots are enabled, in combination with this PR,
then a mismatch will not occur, since the last file to be uploaded
will be a timestamp, which will reference a versioned snapshot
(which will then reference versioned targets).

More details in #171.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>