Add new TUF keyids to keyholder list
Update the keyholders list in the README to include the TUF keyids for the
new key formats introduced in the v5 ceremony.

We continue to list the old TUF keyids for now as they help explain the
continuity from v4 to v5 (it's the same key materials, just different
key formats and keyids).

Signed-off-by: Joshua Lock <jlock@vmware.com>