Don't run cron workflows in forks (#516)
When creating a fork (due to new TUF root veryifing, among others) and
not deleting it after, the cron-scheduled actions would run on the fork
but break due to missing credentials (e.g.,
[1](https://github.com/mihaimaruseac/root-signing/actions/runs/3301257470),
[2](https://github.com/mihaimaruseac/root-signing/actions/runs/3275484644)).

According to https://github.com/orgs/community/discussions/26684,
inserting this `if` statement should prevent the job from running. Did
not insert it in actions that are following the main one since if the
first one does not run none of the others would (if I read the guide
properly). Also, did not add the `if` statement to the other workflow
files since those are not run on cron and should actually run on PRs or
workflow dispatch. Only cron jobs are the ones that need to be disabled.

Signed-off-by: Mihai Maruseac <mihai.maruseac@gmail.com>

Signed-off-by: Mihai Maruseac <mihai.maruseac@gmail.com>