Revert to using origin/main as merge base when running inside actions.
The workflow does not fetch main, only the active branch.
Using origin/main works as that's how it was done before, but it
did cause issues when running the script locally if the user did not
have their origin up-to-date, so we switched to using only main. But
that breaks the PR workflows, this fix changes so in a workflow we
use origin/main, but when running locally it uses main.

Signed-off-by: Fredrik Skogman <kommendorkapten@github.com>