Change timestamp to expire every week, split GHAs
This creates two GHAs to run periodically that will generate timestamps
every Monday and Thursday, and snapshots once a week. The stable
snapshot/timestamp will only run when the repository metadata is
updated.

Preprod to prod will now sync with only a 1 day lag.

Changed the check for a ceremony to be 5 days rather than 7, to give us
2 days for the timestamp GHA to run.

Changed the reusable workflow to pass in flags to enable/disable
timestamp and snapshot.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>