workflows: Replace create-issue with other actions
create-issue had problems with private repos and I decided to rewrite
as it seemd quite fragile. This PR changes root-signing to these new
actions as well.

This does not fix any immediate issues in root-signing but the new
actions should be more maintainable and robust. If this is merged
I intend to remove create-issue action completely.

* The produced comment is not 1:1 identical but very close:
  https://github.com/jku/create-issue-tester/issues/17
* Functionality in root-signing should not change in any way
* The "actions: 'read'" lines included in this commit are not
  really required in this repository, they are only informational

Fixes #1082

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>