Change GCP service account used to manage prod GCS bucket
This new account has the minimum set of roles needed to manage the GCS
bucket and invalidate the CDN cache. It also manages a new KMS signing
key, but that is not currently referenced in this repo.

Staging uses the same github-actions@ service account. This will be
changed once we use sigstore/root-signing-staging to manage the TUF
repository for staging.

Prod still uses an SA under the sigstore-root-signing GCP project, which
houses the production KMS keys.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>