Tweak git attributes to fix default diff behaviour
PR #588 introduced a change to the gitattributes to prevent autocrlf from
changing line endings of files which are digested or size checked (see issue

Telling git to treat all files in the TUF repository as binaries makes the
default git behaviour less useful. While it can be fixed by passing --text,
we can instead modify our gitattributes to treat files in the TUF repository
as LF line-endings only. This change will ensure files are committed with LF
line endings and do not have the line-endings converted on checkout from the
index.

Signed-off-by: Joshua Lock <joshuagloe@gmail.com>