Trigger preproduction sync only after merge to main (#1170)
In #916, we restructured the sync scripts. As part of this, the sync to
preprod after a ceremony completed would occur from the ceremony branch.
We only allow workload impersonation (which is needed to push to GCS and
update the CDN cache) from main, so this breaks.

To fix this, we simply split the workflow into two: The first triggers
on a push to ceremony and creates a PR to merge to main. After merging
from main and updating the root, we sync all contents from the
repository directory.

I also removed the cron job because I don't think it's needed. Also
updated documentation for post-merge steps.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>