workflows: Remove legacy workflows
Delete legacy (meaning non- tuf-on-ci)
 * timestamp and snapshot workflows
 * GCS uploading workflows
 * signing event management workflows (dealing with "ceremony/" dirs)
 * various verification workflows: now handled by the
   tuf-on-ci tests and custom-test.yml

Remove these outright to prevent accidental changes to metadata from two
different maintenance systems.

If it turns out the migration will not be successful in the timeframe it
has, the PR that contains this commit needs to be reverted so the legacy
system can continue operation.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>