workflows: Do not always required future validity in test
test.yml and test-gcs.yml currently always fail if the repository is not
valid for 3 days (or if root & targets are not valid in 30 days). This
is reasonable for standalone tests but not when tests are run during
publish: As an example publishing should succeed after online signing
even if root is expiring in a few weeks.

Do not require repository to be valid at a future date when tests are
running as reusable workflow (in other words from publish.yml), only
require it to be valid at current time.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>