workflows: Fix valid days check
This is actually never true:
    github.event_name == 'workflow_call'
as the event name in workflow_called workflows is the calling workflows event name.

Use the workflow name to distinguish the cases:
 * schedule: name is the test workflow name, use future validity checks
 * workflow_dispatch: name is the test workflow name, use future validity checks
 * workflow_call from publish.yml: name is calling workflown name, do not use
   future validity checks

Using workflow name for this is a bit of a hack and we should consider
adding an input variable for this: that is a bigger change and should be
tested well though.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>